Move utility func to `utils`
